Introduce autoloader, split up files, add build step to deploy, tag 1.3
