+++ .github/workflows/deploy.yml
@@ -11,6 +11,9 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@main
+      - name: Build
+        run: |
+          composer install --no-dev --no-interaction --no-progress --optimize-autoloader
       - name: WordPress Plugin Deploy
         uses: 10up/action-wordpress-plugin-deploy@stable
         env:
