+++ .phpcs.xml.dist
@@ -105,7 +105,7 @@
 	<rule ref="WordPress.NamingConventions.PrefixAllGlobals">
 		<properties>
 			<property name="prefixes" type="array">
-				<element value="Joost"/>
+				<element value="FewerTags"/>
 			</property>
 		</properties>
 	</rule>
